feat: refactor task management by replacing useTaskOperations with useTasks hook and consolidating utility functions for improved maintainability
