+++ frontend/utils/taskUtils.ts
@@ -3,6 +3,13 @@ import { format, fromUnixTime } from "date-fns";
 import { ja } from "date-fns/locale";
 
 /**
+ * Unified task utilities following AGENTS.md structure
+ * Consolidated task-related utility functions for better maintainability
+ */
+
+// ===== SORTING UTILITIES =====
+
+/**
  * Sort tasks by expiry date
  * - Tasks with no expiry date go to the end
  * - Tasks with expiry date are sorted by date (earliest first)
@@ -26,6 +33,41 @@ export function sortCompletedTasks(tasks: Task[]): Task[] {
   return tasks.sort((a, b) => a.title.localeCompare(b.title, 'ja'));
 }
 
+// ===== FILTERING UTILITIES =====
+
+/**
+ * Get tasks by completion status
+ */
+export function filterTasksByCompletion(tasks: Task[], completed: boolean): Task[] {
+  return tasks.filter(task => task.completed === completed);
+}
+
+/**
+ * Filter expired tasks
+ */
+export function getExpiredTasks(tasks: Task[]): Task[] {
+  const currentTimeInSeconds = Math.floor(Date.now() / 1000);
+  return tasks.filter(task => 
+    task.expires_at && currentTimeInSeconds > task.expires_at
+  );
+}
+
+/**
+ * Filter tasks expiring soon (within specified hours)
+ */
+export function getTasksExpiringSoon(tasks: Task[], withinHours: number = 24): Task[] {
+  const currentTimeInSeconds = Math.floor(Date.now() / 1000);
+  const thresholdTime = currentTimeInSeconds + (withinHours * 60 * 60);
+  
+  return tasks.filter(task => 
+    task.expires_at && 
+    task.expires_at > currentTimeInSeconds && 
+    task.expires_at <= thresholdTime
+  );
+}
+
+// ===== STATUS UTILITIES =====
+
 /**
  * Check if a task is expired
  */
@@ -34,6 +76,71 @@ export function isTaskExpired(task: Task): boolean {
   const currentTimeInSeconds = Math.floor(Date.now() / 1000);
   return currentTimeInSeconds > task.expires_at;
 }
+
+/**
+ * Check if a task is expiring soon
+ */
+export function isTaskExpiringSoon(task: Task, withinHours: number = 24): boolean {
+  if (!task.expires_at) return false;
+  const currentTimeInSeconds = Math.floor(Date.now() / 1000);
+  const thresholdTime = currentTimeInSeconds + (withinHours * 60 * 60);
+  return task.expires_at > currentTimeInSeconds && task.expires_at <= thresholdTime;
+}
+
+/**
+ * Get comprehensive task status for display
+ */
+export function getTaskStatus(task: Task): {
+  isExpired: boolean;
+  isExpiringSoon: boolean;
+  isCompleted: boolean;
+  statusText: string;
+  variant: 'default' | 'secondary' | 'destructive' | 'warning';
+} {
+  const isExpired = isTaskExpired(task);
+  const isExpiringSoon = isTaskExpiringSoon(task);
+  const isCompleted = task.completed;
+
+  if (isCompleted) {
+    return {
+      isExpired,
+      isExpiringSoon,
+      isCompleted,
+      statusText: '完了',
+      variant: 'secondary'
+    };
+  }
+
+  if (isExpired) {
+    return {
+      isExpired,
+      isExpiringSoon,
+      isCompleted,
+      statusText: '期限切れ',
+      variant: 'destructive'
+    };
+  }
+
+  if (isExpiringSoon) {
+    return {
+      isExpired,
+      isExpiringSoon,
+      isCompleted,
+      statusText: 'まもなく期限',
+      variant: 'warning'
+    };
+  }
+
+  return {
+    isExpired,
+    isExpiringSoon,
+    isCompleted,
+    statusText: 'アクティブ',
+    variant: 'default'
+  };
+}
+
+// ===== FORMATTING UTILITIES =====
 
 /**
  * Format task expiry date for display
@@ -45,49 +152,56 @@ export function formatTaskExpiry(task: Task, formatString: string = 'yyyy年M月
 }
 
 /**
- * Get task status for display
+ * Get relative time description for task expiry
  */
-export function getTaskStatus(task: Task): {
-  isExpired: boolean;
-  isCompleted: boolean;
-  statusText: string;
-  variant: 'default' | 'secondary' | 'destructive';
-} {
-  const isExpired = isTaskExpired(task);
-  const isCompleted = task.completed;
-
-  if (isCompleted) {
-    return {
-      isExpired,
-      isCompleted,
-      statusText: '完了',
-      variant: 'secondary'
-    };
+export function getTaskExpiryRelativeTime(task: Task): string | null {
+  if (!task.expires_at) return null;
+  
+  const currentTimeInSeconds = Math.floor(Date.now() / 1000);
+  const diffInSeconds = task.expires_at - currentTimeInSeconds;
+  
+  if (diffInSeconds < 0) {
+    return '期限切れ';
   }
-
-  if (isExpired) {
-    return {
-      isExpired,
-      isCompleted,
-      statusText: '期限切れ',
-      variant: 'destructive'
-    };
+  
+  const diffInMinutes = Math.floor(diffInSeconds / 60);
+  const diffInHours = Math.floor(diffInMinutes / 60);
+  const diffInDays = Math.floor(diffInHours / 24);
+  
+  if (diffInDays > 0) {
+    return `${diffInDays}日後`;
+  } else if (diffInHours > 0) {
+    return `${diffInHours}時間後`;
+  } else if (diffInMinutes > 0) {
+    return `${diffInMinutes}分後`;
+  } else {
+    return 'まもなく';
   }
+}
 
+// ===== STATISTICS UTILITIES =====
+
+/**
+ * Get task statistics
+ */
+export function getTaskStatistics(tasks: Task[]) {
+  const total = tasks.length;
+  const completed = tasks.filter(task => task.completed).length;
+  const active = total - completed;
+  const expired = getExpiredTasks(tasks.filter(task => !task.completed)).length;
+  const expiringSoon = getTasksExpiringSoon(tasks.filter(task => !task.completed)).length;
+  
   return {
-    isExpired,
-    isCompleted,
-    statusText: 'アクティブ',
-    variant: 'default'
+    total,
+    completed,
+    active,
+    expired,
+    expiringSoon,
+    completionRate: total > 0 ? Math.round((completed / total) * 100) : 0
   };
 }
 
-/**
- * Get tasks by completion status
- */
-export function filterTasksByCompletion(tasks: Task[], completed: boolean): Task[] {
-  return tasks.filter(task => task.completed === completed);
-}
+// ===== LEGACY FUNCTIONS (DEPRECATED) =====
 
 /**
  * Format expiry date for display (legacy - use formatTaskExpiry instead)
